+++ src/convert/v1ToV2.ts
@@ -7,21 +7,21 @@ import { FigTreeEvaluator } from '../FigTreeEvaluator'
 import { isObject, standardiseOperatorName } from '../helpers'
 import { EvaluatorNode, OperatorNode } from '../types'
 
-export const convertV1ToV2 = async (expression: EvaluatorNode, figTree: FigTreeEvaluator) => {
+export const convertV1ToV2 = (expression: EvaluatorNode, figTree: FigTreeEvaluator) => {
   const operators = figTree.getOperators()
 
-  const v1ToV2 = async (expression: EvaluatorNode) => {
+  const v1ToV2 = (expression: EvaluatorNode) => {
     if (!isObject(expression)) return expression
 
     const newExpression = isV1Node(expression)
-      ? await (async () => {
+      ? (() => {
           const { operator } = expression
           const standardisedOpName = standardiseOperatorName(operator)
           const operatorRef = operators.find(
             (op) => op.name === standardisedOpName || op.aliases.includes(standardisedOpName)
           )
           if (operatorRef) {
-            const result = await operatorRef.parseChildren(expression, figTree.getConfig())
+            const result = operatorRef.parseChildren(expression, figTree.getConfig())
             delete result.children
             const preferredOpName = operatorRef.aliases[0]
             result.operator = preferredOpName
@@ -35,10 +35,8 @@ export const convertV1ToV2 = async (expression: EvaluatorNode, figTree: FigTreeE
     for (const [key, value] of Object.entries(newExpression as object)) {
       const modifiedKey = key === 'type' ? 'outputType' : key
       if (Array.isArray(value)) {
-        const newArray = value.map((val) => v1ToV2(val))
-        const resolved = await Promise.all(newArray)
-        outputExpression[modifiedKey] = resolved
-      } else outputExpression[modifiedKey] = await v1ToV2(value)
+        outputExpression[modifiedKey] = value.map((val) => v1ToV2(val))
+      } else outputExpression[modifiedKey] = v1ToV2(value)
     }
 
     return outputExpression